+++ src/header_crypto/decrypt.rs
@@ -1,12 +1,15 @@
 use crate::header_crypto::encrypt::EncrypterHalf;
-use crate::header_crypto::Decrypter;
+use crate::header_crypto::{
+    ClientHeader, ServerHeader, CLIENT_HEADER_LENGTH, SERVER_HEADER_LENGTH,
+};
 use crate::SESSION_KEY_LENGTH;
+use std::io::Read;
 
 /// Decryption part of a [`HeaderCrypto`](crate::header_crypto::HeaderCrypto).
 ///
 /// Intended to be kept with the reader half of a connection.
 ///
-/// Use the [`Decrypter`] functions to decrypt.
+/// Use the [`DecrypterHalf`] functions to decrypt.
 #[derive(Debug)]
 pub struct DecrypterHalf {
     pub(crate) session_key: [u8; SESSION_KEY_LENGTH as usize],
@@ -14,13 +17,13 @@ pub struct DecrypterHalf {
     pub(crate) previous_value: u8,
 }
 
-impl Decrypter for DecrypterHalf {
-    /// Use either [the client](Decrypter::read_and_decrypt_client_header)
-    /// or [the server](Decrypter::read_and_decrypt_server_header)
+impl DecrypterHalf {
+    /// Use either [the client](DecrypterHalf::read_and_decrypt_client_header)
+    /// or [the server](DecrypterHalf::read_and_decrypt_server_header)
     /// [`Read`](std::io::Read) functions, or
-    /// [the client](Decrypter::decrypt_client_header)
-    /// or [the server](Decrypter::decrypt_server_header) array functions.
-    fn decrypt(&mut self, data: &mut [u8]) {
+    /// [the client](DecrypterHalf::decrypt_client_header)
+    /// or [the server](DecrypterHalf::decrypt_server_header) array functions.
+    pub fn decrypt(&mut self, data: &mut [u8]) {
         decrypt(
             data,
             &self.session_key,
@@ -28,9 +31,67 @@ impl Decrypter for DecrypterHalf {
             &mut self.previous_value,
         );
     }
-}
 
-impl DecrypterHalf {
+    /// [`Read`](std::io::Read) wrapper for [`DecrypterHalf::decrypt_server_header`].
+    ///
+    /// # Errors
+    ///
+    /// Has the same errors as [`std::io::Read::read_exact`].
+    pub fn read_and_decrypt_server_header<R: Read>(
+        &mut self,
+        reader: &mut R,
+    ) -> std::io::Result<ServerHeader> {
+        let mut buf = [0_u8; SERVER_HEADER_LENGTH as usize];
+        reader.read_exact(&mut buf)?;
+
+        Ok(self.decrypt_server_header(buf))
+    }
+
+    /// [`Read`](std::io::Read) wrapper for [`DecrypterHalf::decrypt_client_header`].
+    ///
+    /// # Errors
+    ///
+    /// Has the same errors as [`std::io::Read::read_exact`].
+    pub fn read_and_decrypt_client_header<R: Read>(
+        &mut self,
+        reader: &mut R,
+    ) -> std::io::Result<ClientHeader> {
+        let mut buf = [0_u8; CLIENT_HEADER_LENGTH as usize];
+        reader.read_exact(&mut buf)?;
+
+        Ok(self.decrypt_client_header(buf))
+    }
+
+    /// Convenience function for decrypting server headers.
+    ///
+    /// Prefer this over directly using [`DecrypterHalf::decrypt`].
+    pub fn decrypt_server_header(
+        &mut self,
+        mut data: [u8; SERVER_HEADER_LENGTH as usize],
+    ) -> ServerHeader {
+        self.decrypt(&mut data);
+
+        let size = u16::from_be_bytes([data[0], data[1]]);
+        let opcode = u16::from_le_bytes([data[2], data[3]]);
+
+        ServerHeader { size, opcode }
+    }
+
+    /// Convenience function for decrypting client headers.
+    ///
+    /// Prefer this over directly using [`DecrypterHalf::decrypt`].
+    pub fn decrypt_client_header(
+        &mut self,
+        mut data: [u8; CLIENT_HEADER_LENGTH as usize],
+    ) -> ClientHeader {
+        self.decrypt(&mut data);
+
+        let size: u16 = u16::from_be_bytes([data[0], data[1]]);
+        let opcode: u32 = u32::from_le_bytes([data[2], data[3], data[4], data[5]]);
+
+        ClientHeader { size, opcode }
+    }
+
     /// Tests whether both halves originate from the same
     /// [`HeaderCrypto`](crate::header_crypto::HeaderCrypto)
     /// and can be [`EncrypterHalf::unsplit`].
@@ -44,14 +105,11 @@ impl DecrypterHalf {
         Self {
             session_key,
             index: 0,
-            previous_value: 0
+            previous_value: 0,
         }
     }
 }
 
-// Separate function to prevent duplicating logic in full and half versions.
-// The half version isn't used in order to allow the full version to only have
-// one session key instead of 2, saving 32 bytes.
 pub(crate) fn decrypt(
     data: &mut [u8],
     session_key: &[u8; SESSION_KEY_LENGTH as usize],